Add missing s-ze_t, and a size_t to an int.

+++ libntp/msyslog.c
@@ -287,7 +287,7 @@ msnprintf(
 	)
 {
 	va_list	ap;
-	size_t	rc;
+	int	rc;
 
 	va_start(ap, fmt);
 	rc = mvsnprintf(buf, bufsiz, fmt, ap);
